Convert HitEffectiveLevelModifier level from Double to Integer

+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/combat/hit/HitEffectiveLevelModifier.kt
@@ -12,5 +12,5 @@ import world.gregs.voidps.engine.event.Event
 data class HitEffectiveLevelModifier(
     val skill: Skill,
     val accuracy: Boolean,
-    var level: Double
+    var level: Int
 ) : Event
